Script update for Watson Discovery 4.0.6

+++ discovery-data/2.2.0/post-restore.sh
@@ -59,7 +59,6 @@ do
   fi
 done
 
-brlog "INFO" "Restore tenants information"
 WD_VERSION=${WD_VERSION:-`get_version`}
 PG_POD=""
 
@@ -72,34 +71,34 @@ if [ `compare_version "${WD_VERSION}" "4.0.0"` -ge 0 ] ; then
   run_pg_job
   PG_POD=${POD}
 else
-  for POD in `oc get pods ${OC_ARGS} -o jsonpath='{.items[*].metadata.name}' -l tenant=${TENANT_NAME},component=stolon-keeper` ; do
-    if oc logs ${OC_ARGS} --since=30s ${POD} | grep 'our db requested role is master' > /dev/null ; then
-      PG_POD=${POD}
+  PG_POD=$(get_primary_pg_pod)
+fi
+
+if require_tenant_backup ; then
+
+  brlog "INFO" "Restore tenants information"
+
+  for tenants_file in `ls -t tmp_wd_tenants_*.txt` ; do
+    if [ -n "`cat ${tenants_file}`" ] ; then
+      kube_cp_from_local ${PG_POD} "${tenants_file}" "/tmp/tenants" ${OC_ARGS}
+      fetch_cmd_result ${PG_POD} 'export PGUSER=${PGUSER:-$STKEEPER_PG_SU_USERNAME} && \
+        export PGPASSWORD=${PGPASSWORD:-$STKEEPER_PG_SU_PASSWORD} && \
+        export PGHOST=${PGHOST:-$HOSTNAME} && \
+        while ! psql -d dadmin -t -c "SELECT * FROM tenants" &> /dev/null; do sleep 10; done && \
+        if ! psql -d dadmin -t -c "SELECT * FROM tenants;" | grep "default" > /dev/null ; then \
+          psql -d dadmin -c "TRUNCATE tenants" && \
+          psql -d dadmin -c "\COPY tenants FROM '"'"'/tmp/tenants'"'"'" ;\
+        else\
+          echo "COPY" ;\
+        fi&& \
+        rm -f /tmp/tenants' ${OC_ARGS} | grep COPY >& /dev/null
     fi
   done
 fi
-
-for tenants_file in `ls -t tmp_wd_tenants_*.txt` ; do
-  if [ -n "`cat ${tenants_file}`" ] ; then
-    kube_cp_from_local ${PG_POD} "${tenants_file}" "/tmp/tenants" ${OC_ARGS}
-    fetch_cmd_result ${PG_POD} 'export PGUSER=${PGUSER:-$STKEEPER_PG_SU_USERNAME} && \
-      export PGPASSWORD=${PGPASSWORD:-$STKEEPER_PG_SU_PASSWORD} && \
-      export PGHOST=${PGHOST:-$HOSTNAME} && \
-      while ! psql -d dadmin -t -c "SELECT * FROM tenants" &> /dev/null; do sleep 10; done && \
-      if ! psql -d dadmin -t -c "SELECT * FROM tenants;" | grep "default" > /dev/null ; then \
-        psql -d dadmin -c "TRUNCATE tenants" && \
-        psql -d dadmin -c "\COPY tenants FROM '"'"'/tmp/tenants'"'"'" ;\
-      else\
-        echo "COPY" ;\
-      fi&& \
-      rm -f /tmp/tenants' ${OC_ARGS} | grep COPY >& /dev/null
-  fi
-done
-
 ## End restore tenants
 
 ## Set default as tenant ID
-if [ `compare_version "${WD_VERSION}" "2.2.1"` -ge 0 ] && [ `compare_version "${BACKUP_FILE_VERSION}" "2.2.0"` -le 0 ] ; then
+if [ `compare_version "${WD_VERSION}" "2.2.1"` -ge 0 ] && [ `compare_version "${WD_VERSION}" "4.0.5"` -le 0 ] && [ `compare_version "${BACKUP_FILE_VERSION}" "2.2.0"` -le 0 ] ; then
   fetch_cmd_result ${PG_POD} 'export PGUSER=${PGUSER:-$STKEEPER_PG_SU_USERNAME} && \
       export PGPASSWORD=${PGPASSWORD:-$STKEEPER_PG_SU_PASSWORD} && \
       export PGHOST=${PGHOST:-$HOSTNAME} && \
@@ -107,6 +106,8 @@ if [ `compare_version "${WD_VERSION}" "2.2.1"` -ge 0 ] && [ `compare_version "${
       psql -d dadmin -c "UPDATE datasets SET tenant_id = '"'default'"' ;" ' ${OC_ARGS}
 fi
 ## End set default
+
+## Update ranker version to run retrain
 fetch_cmd_result ${PG_POD} 'export PGUSER=${PGUSER:-$STKEEPER_PG_SU_USERNAME} && \
       export PGPASSWORD=${PGPASSWORD:-$STKEEPER_PG_SU_PASSWORD} && \
       export PGHOST=${PGHOST:-$HOSTNAME} && \
@@ -114,7 +115,6 @@ fetch_cmd_result ${PG_POD} 'export PGUSER=${PGUSER:-$STKEEPER_PG_SU_USERNAME} &&
       psql -d ranker_training -c "WITH T AS (SELECT DISTINCT ON (training_set_id) training_job_id from jobs ORDER BY training_set_id, data_version DESC) UPDATE jobs SET state = '"'"'INVALIDATE_RANKER'"'"' where training_job_id IN (select training_job_id from T);"
       ' ${OC_ARGS}
 
-## Update ranker version to run retrain
 
 ## End update ranker
